model db

+++ twitoff/db_model.py
@@ -16,10 +16,10 @@ class User(DB.Model):
 
 
 class Tweet(DB.Model):
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     tweet = DB.Column(DB.String(280), unique=True, nullable=False)
     embedding = DB.Column(DB.PickleType, nullable=False)
-    user_id = DB.Column(DB.Integer, DB.ForeignKey('user.id'), nullable=False)
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'), nullable=False)
     user = DB.relationship('User', backref=DB.backref('tweet', lazy=True))
 
     def __repr__(self):
